move testdata to fixtures
