+++ tests/lib/helpers/problemGenerator.js
@@ -1,0 +1,8 @@
+module.exports = function (Rule, location, message, fixProps = {}) {
+    return {
+        ...Rule.meta,
+        location,
+        message,
+        ...fixProps,
+    };
+};
